Move format as a method on the FixedDateRange type
