+++ lib/date/index.ts
@@ -1,4 +1,3 @@
 export * from "./dayjs"
 export * from "./Date"
 export * from "./DateRange"
-export * from "./Formatting"
